statistics: collect colors in a StatsTheme class
To enable rudimentary theming, collect all colors in a new theme class.
The class has to be passed down to the various items.

In general the items save a reference to the theme in the constructor.
Alternatively, they might also just query the StatsView everytime they
need to access a color.
For now, it's hard the say what is preferred: a reference per item or a
function call per invokation?

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>